refactor: game 객체 제거
